move qualified column set to where it is used
