+++ snuba/query/data_source/__init__.py
@@ -1,32 +1,6 @@
 from abc import ABC, abstractmethod
-from typing import Mapping, MutableSequence
 
-from snuba.clickhouse.columns import Column
-from snuba.utils.schemas import ColumnSet, SchemaModifiers, WildcardColumn
-
-
-class QualifiedColumnSet(ColumnSet):
-    """
-    Works like a Columnset but it represent a list of columns
-    coming from different tables (like the ones we would use in
-    a join).
-    The main difference is that this class keeps track of the
-    structure and to which table each column belongs to.
-    """
-
-    def __init__(self, column_sets: Mapping[str, ColumnSet]) -> None:
-        columns: MutableSequence[Column[SchemaModifiers]] = []
-
-        for alias, column_set in column_sets.items():
-            for column in column_set.columns:
-                if isinstance(column, WildcardColumn):
-                    columns.append(
-                        WildcardColumn(f"{alias}.{column.name}", column.type)
-                    )
-                else:
-                    columns.append(Column(f"{alias}.{column.name}", column.type))
-
-        super().__init__(columns)
+from snuba.utils.schemas import ColumnSet
 
 
 class DataSource(ABC):
